+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
-
+- Crossref lookups via Habanero now comply with the "be-nice" policy
+- Removed `UnboundLocalError` from error processing for reference validation
 
 ## [0.3.0] - 2017-10-09
 ### Added